feat: Change resources endpoints (`img`, `resource`).
