+++ config/default.json
@@ -281,6 +281,7 @@
         "adempiere/dictionary/search-info",
         "adempiere/general-ledger",
         "adempiere/general-ledger/accouting-combination",
+        "adempiere/img",
         "adempiere/material-management",
         "adempiere/material-management/product-attribute",
         "adempiere/material-management/warehouse-locator",
@@ -331,7 +332,8 @@
         "adempiere/order",
         "adempiere/payment",
         "adempiere/product",
-        "adempiere/record-management"
+        "adempiere/record-management",
+        "adempiere/resource"
       ],
       "images": {
         "baseUrl": "http://localhost",
